Add a dispatcher binary
